requirements.txt
claude-agent-sdk>=0.1.0
alpaca-py>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
pytz
discord.py>=2.4.0
websockets>=12.0
matplotlib>=3.8.0
playwright>=1.50.0
jinja2>=3.1.0
fastapi>=0.115.0
uvicorn>=0.30.0
